refactor: Move some ELF-specific code from layout.rs to elf.rs
Issue #1538
