+++ libwild/src/loongarch64.rs
@@ -86,7 +86,7 @@ impl crate::elf_arch::ElfArch for LoongArch64 {
         layout.tls_start_address()
     }
 
-    fn get_property_class(_property_type: u32) -> Option<crate::layout::PropertyClass> {
+    fn get_property_class(_property_type: u32) -> Option<crate::elf::PropertyClass> {
         None
     }
